+++ pydigitalstrom/devices/blind.py
@@ -4,7 +4,10 @@ from pydigitalstrom.exceptions import DSParameterException
 
 
 class DSBlind(DSTerminal):
-    URL_STATUS = '/json/device/getOutputValue?dsid={id}&offset=0'
+    URL_UPDATE = '/json/device/getOutputValue?dsid={id}&offset=0'
+    URL_ACTION_STOP = '/json/device/callScene?dsid={id}&sceneNumber=15'
+    URL_ACTION_UP = '/json/device/callScene?dsid={id}&sceneNumber=14'
+    URL_ACTION_DOWN = '/json/device/callScene?dsid={id}&sceneNumber=13'
     URL_SET_POSITION = '/json/device/setValue?dsid={id}&value={position}'
 
     def __init__(self, *args, **kwargs):
@@ -18,8 +21,17 @@ class DSBlind(DSTerminal):
         if position < 0 or position > 255:
             raise DSParameterException('blind position must be an integer between 0 and 255')
 
-        self.request(self.URL_SET_POSITION, position=position)
+        self.request(url=self.URL_SET_POSITION, position=position)
         self._position = position
 
+    def stop(self):
+        self.request(url=self.URL_ACTION_STOP, check_result=False)
+
+    def move_up(self):
+        self.request(url=self.URL_ACTION_UP, check_result=False)
+
+    def move_down(self):
+        self.request(url=self.URL_ACTION_DOWN, check_result=False)
+
     def update(self):
-        self._position = self.request(self.URL_STATUS)['value']
+        self._position = self.request(url=self.URL_UPDATE)['value']