+++ pydigitalstrom/devices/server.py
@@ -4,9 +4,19 @@ from pydigitalstrom.devices.base import DSDevice
 
 
 class DSServer(DSDevice):
+    URL_SYSTEM_INFO = '/json/system/version'
+
     def __init__(self, client: DSClient, data: dict, *args, **kwargs):
         super().__init__(client=client, *args, **kwargs)
         self._data = data
+
+    @property
+    def name(self):
+        return 'server'
+
+    @property
+    def unique_id(self):
+        return self._data['MachineID']
 
     def get_data(self):
         """
@@ -21,4 +31,4 @@ class DSServer(DSDevice):
         """
         update server meta data
         """
-        self._data = self.request(url=self._client.URL_SYSTEM_INFO)
+        self._data = self.request(url=self.URL_SYSTEM_INFO)